add Group.methods.leave, user is removed from groups when his account is deleted
