+++ website/src/controllers/groups.js
@@ -235,6 +235,7 @@ api.update = function(req, res, next) {
   });
 }
 
+// TODO remove from api object?
 api.attachGroup = function(req, res, next) {
   var gid = req.params.gid;
   var q = (gid == 'party') ? Group.findOne({type: 'party', members: {'$in': [res.locals.user._id]}}) : Group.findById(gid);
@@ -243,7 +244,7 @@ api.attachGroup = function(req, res, next) {
     if(!group) return res.json(404, {err: "Group not found"});
     res.locals.group = group;
     next();
-  })
+  });
 }
 
 api.getChat = function(req, res, next) {
@@ -487,92 +488,18 @@ api.join = function(req, res, next) {
 }
 
 api.leave = function(req, res, next) {
-  var user = res.locals.user,
-    group = res.locals.group;
+  var user = res.locals.user;
+  var group = res.locals.group;
+
   // When removing the user from challenges, should we keep the tasks?
   var keep = (/^remove-all/i).test(req.query.keep) ? 'remove-all' : 'keep-all';
-  async.parallel([
-    // Remove active quest from user if they're leaving the party
-    function(cb){
-      if (group.type != 'party') return cb(null,{},1);
-      user.party.quest = Group.cleanQuestProgress();
-      user.save(cb);
-    },
-    // Remove user from group challenges
-    function(cb){
-      async.waterfall([
-        // Find relevant challenges
-        function(cb2) {
-          Challenge.find({
-            _id: {$in: user.challenges}, // Challenges I am in
-            group: group._id // that belong to the group I am leaving
-          }, cb2);
-        },
-        // Update each challenge
-        function(challenges, cb2) {
-          Challenge.update(
-            {_id:{$in: _.pluck(challenges, '_id')}},
-            {$pull:{members:user._id}},
-            {multi: true},
-            function(err) {
-             cb2(err, challenges); // pass `challenges` above to cb
-            }
-          );
-        },
-        // Unlink the challenge tasks from user
-        function(challenges, cb2) {
-          async.waterfall(challenges.map(function(chal) {
-            return function(cb3) {
-              var i = user.challenges.indexOf(chal._id)
-              if (~i) user.challenges.splice(i,1);
-              user.unlink({cid:chal._id, keep:keep}, cb3);
-            }
-          }), cb2);
-        }
-      ], cb);
-    },
-    // Update the group
-    function(cb){
-      // If user is the last one in group and group is private, delete it
-      if(group.members.length === 1 && (
-          group.type === 'party' ||
-          (group.type === 'guild' && group.privacy === 'private')
-      )){
-        // TODO remove invitations to this group
-        Group.remove({
-          _id: group._id
-        }, cb);
-      }else{ // otherwise just remove a member
-        var update = {$pull:{members:user._id}};
-        if (group.type == 'party' && group.quest.key){
-          update['$unset'] = {};
-          update['$unset']['quest.members.' + user._id] = 1;
-        }
-        // FIXME do we want to remove the group `if group.members.length == 0` ? (well, 1 since the update hasn't gone through yet)
-        if (group.members.length > 1) {
-          var seniorMember = _.find(group.members, function (m) {return m != user._id});
-          // If the leader is leaving (or if the leader previously left, and this wasn't accounted for)
-          var leader = group.leader;
-          if (leader == user._id || !~group.members.indexOf(leader)) {
-            update['$set'] = update['$set'] || {};
-            update['$set'].leader = seniorMember;
-          }
-          leader = group.quest && group.quest.leader;
-          if (leader && (leader == user._id || !~group.members.indexOf(leader))) {
-            update['$set'] = update['$set'] || {};
-            update['$set']['quest.leader'] = seniorMember;
-          }
-        }
-        update['$inc'] = {memberCount: -1};
-        Group.update({_id:group._id},update,cb);
-      }
-    }
-  ],function(err){
+
+  group.leave(user, keep, function(err){
     if (err) return next(err);
     user = group = keep = null;
     return res.send(204);
-  })
-}
+  });
+};
 
 var inviteByUUIDs = function(uuids, group, req, res, next){
   async.each(uuids, function(uuid, cb){
